https://issues.apache.org/jira/browse/AMQ-3374 - long kahadb tx - parse long
git-svn-id: https://svn.apache.org/repos/asf/activemq/trunk@1138449 13f79535-47bb-0310-9956-ffa450edef68

+++ kahadb/src/main/java/org/apache/kahadb/page/Transaction.java
@@ -91,7 +91,7 @@ public class Transaction implements Iterable<Page> {
     // List of pages freed in this transaction
     private final SequenceSet freeList = new SequenceSet();
 
-    private long maxTransactionSize = Integer.parseInt(System.getProperty("maxKahaDBTxSize", "" + 10485760));
+    private long maxTransactionSize = Long.parseLong(System.getProperty("maxKahaDBTxSize", "" + 10485760));
 
     private long size = 0;
